requirements.txt
Flask==0.12.2
gunicorn==19.7.1
pandas==0.22.0
PyYAML==3.12
validators==0.12.2
